+++ SwiftMessages/WindowViewController.swift
@@ -11,26 +11,33 @@ import UIKit
 class WindowViewController: UIViewController
 {
     fileprivate var window: UIWindow?
-    fileprivate var appWindow: UIWindow?
 
     let windowLevel: UIWindowLevel
-    var statusBarStyle: UIStatusBarStyle?
-
-    init(windowLevel: UIWindowLevel = UIWindowLevelNormal)
+    let config: SwiftMessages.Config
+    
+    override var shouldAutorotate: Bool {
+        return config.shouldAutorotate
+    }
+    
+    init(windowLevel: UIWindowLevel = UIWindowLevelNormal, config: SwiftMessages.Config)
     {
         self.windowLevel = windowLevel
+        self.config = config
         let window = PassthroughWindow(frame: UIScreen.main.bounds)
         self.window = window
-        appWindow = UIApplication.shared.keyWindow
         super.init(nibName: nil, bundle: nil)
         self.view = PassthroughView()
         window.rootViewController = self
         window.windowLevel = windowLevel
     }
-
-    func install() {
+    
+    func install(becomeKey: Bool) {
         guard let window = window else { return }
-        window.makeKeyAndVisible()
+        if becomeKey {
+            window.makeKeyAndVisible()            
+        } else {
+            window.isHidden = false
+        }
     }
 
     func uninstall() {
@@ -43,17 +50,10 @@ class WindowViewController: UIViewController
     }
     
     override public var preferredStatusBarStyle: UIStatusBarStyle {
-        return statusBarStyle ?? UIApplication.shared.statusBarStyle
+        return config.preferredStatusBarStyle ?? UIApplication.shared.statusBarStyle
     }
-
+    
     override var prefersStatusBarHidden: Bool {
-        return appWindow?.rootViewController?.prefersStatusBarHidden ?? UIApplication.shared.isStatusBarHidden
-    }
-
-    override func didRotate(from fromInterfaceOrientation: UIInterfaceOrientation) {
-        super.didRotate(from: fromInterfaceOrientation)
-        DispatchQueue.main.async {
-            self.setNeedsStatusBarAppearanceUpdate()
-        }
+        return UIApplication.shared.isStatusBarHidden
     }
 }